move shared validators to shared
