+++ packages/back-end/types/feature.d.ts
@@ -29,7 +29,7 @@ export {
   NamespaceValue,
   SavedGroupTargeting,
   FeaturePrerequisite,
-} from "back-end/src/validators/shared";
+} from "shared/src/validators/shared";
 
 export type SchemaField = z.infer<typeof simpleSchemaFieldValidator>;
 export type SimpleSchema = z.infer<typeof simpleSchemaValidator>;
