+++ packages/shared/src/validators/shared.ts
